testing/web-platform/tests/service-workers/service-worker/update-bytecheck.https.html

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src="/resources/testharness.js"></script>
<script src="resources/testharness-helpers.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>

promise_test(function(t) {
    var script = 'resources/bytecheck-worker.py' +
                 '?main=default' +
                 '&imported=default';
    var scope = 'resources/blank.html';

    var swr, sw;
    return Promise.resolve()
        .then(() => service_worker_unregister_and_register(t, script, scope))
        .then(registration => swr = registration)

        .then(() => wait_for_update(t, swr))
        .then(worker => sw = worker)

        .then(() => wait_for_state(t, sw, 'activated'))
        .then(() => assert_array_equals([swr.active,
                                         swr.waiting,
                                         swr.installing],
                                        [sw,
                                         null,
                                         null]))

        .then(() => swr.update())
        .then(() => assert_array_equals([swr.active,
                                         swr.waiting,
                                         swr.installing],
                                        [sw,
                                         null,
                                         null]))

        .then(() => service_worker_unregister_and_done(t, scope));
}, "Test with (main: same, imported: same)");

promise_test(function(t) {
    var script = 'resources/bytecheck-worker.py' +
                 '?main=default' +
                 '&imported=time';
    var scope = 'resources/blank.html';

    var swr, sw;
    return Promise.resolve()
        .then(() => service_worker_unregister_and_register(t, script, scope))
        .then(registration => swr = registration)

        .then(() => wait_for_update(t, swr))
        .then(worker => sw = worker)

        .then(() => wait_for_state(t, sw, 'activated'))
        .then(() => assert_array_equals([swr.active,
                                         swr.waiting,
                                         swr.installing],
                                        [sw,
                                         null,
                                         null]))

        .then(() => swr.update())
        .then(() => wait_for_update(t, swr))

        .then(() => service_worker_unregister_and_done(t, scope));
}, "Test with (main: same, imported: different)");

promise_test(function(t) {
    var script = 'resources/bytecheck-worker.py' +
                 '?main=time' +
                 '&imported=default';
    var scope = 'resources/blank.html';

    var swr, sw;
    return Promise.resolve()
        .then(() => service_worker_unregister_and_register(t, script, scope))
        .then(registration => swr = registration)

        .then(() => wait_for_update(t, swr))
        .then(worker => sw = worker)

        .then(() => wait_for_state(t, sw, 'activated'))
        .then(() => assert_array_equals([swr.active,
                                         swr.waiting,
                                         swr.installing],
                                        [sw,
                                         null,
                                         null]))

        .then(() => swr.update())
        .then(() => wait_for_update(t, swr))

        .then(() => service_worker_unregister_and_done(t, scope));
}, "Test with (main: different, imported: same)");

promise_test(function(t) {
    var script = 'resources/bytecheck-worker.py' +
                 '?main=time' +
                 '&imported=time';
    var scope = 'resources/blank.html';

    var swr, sw;
    return Promise.resolve()
        .then(() => service_worker_unregister_and_register(t, script, scope))
        .then(registration => swr = registration)

        .then(() => wait_for_update(t, swr))
        .then(worker => sw = worker)

        .then(() => wait_for_state(t, sw, 'activated'))
        .then(() => assert_array_equals([swr.active,
                                         swr.waiting,
                                         swr.installing],
                                        [sw,
                                         null,
                                         null]))

        .then(() => swr.update())
        .then(() => wait_for_update(t, swr))

        .then(() => service_worker_unregister_and_done(t, scope));
}, "Test with (main: different, imported: different)");

</script>
